requirements.txt
# Core Libraries
torch>=2.0.0
torchvision>=0.15.0
# torchaudio # Add if using torchaudio features
numpy>=1.20.0
pillow>=8.0.0

# Data Handling & Processing
librosa>=0.8.0
PyYAML>=5.4 # For loading YAML config files

# Models (if using Hugging Face)
transformers>=4.10.0 # For SwinTransformerBackbone
# timm # Optional, if using timm models as backbones

# Utilities
tqdm # For progress bars (optional, but nice)
matplotlib # For plotting (optional, e.g., saving loss curves)

# Specify versions as needed, e.g., torch==1.13.1